requirements.txt
bcrypt==4.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.4.4
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.31
mysql-connector-python-rf==2.2.2
protobuf==3.20.1
psycopg2==2.9.5
psycopg2-binary==2.9.5
PyJWT==2.6.0
pypi==2.1
requests==2.28.1
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.10.0
